#259 convert tests to typescript - moved setup of context out of indiviudal test files into new file base.test.ts as it was not working like intended before
