+++ binocular-backend/test/paginator.test.ts
@@ -4,10 +4,9 @@ import _ from 'lodash';
 import chai from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
-
 chai.use(sinonChai);
-
 import Paginator from '../utils/paginator';
+import './base.test.ts';
 
 const expect = chai.expect;
