+++ binocular-backend/test/vcsIndexer.test.ts
@@ -28,32 +28,15 @@ import conf from '../utils/config.js';
 import ctx from '../utils/context';
 import GitHubUrlProvider from '../url-providers/GitHubUrlProvider';
 import VcsIndexer from '../indexers/vcs';
-import path from 'path';
 import { checkOwnedLines, expectExamples, getAllEntriesInCollection, getAllRelevantCollections } from './helper/utils.ts';
 import { alice, bob, file1, file1Change1, file1Change2, file2 } from './helper/git/gitTestData.ts';
 import GitIndexer from '../indexers/vcs/GitIndexer';
 import Model from '../models/Model.ts';
 import Connection from '../models/Connection.ts';
-const indexerOptions = {
-  backend: true,
-  frontend: false,
-  open: false,
-  clean: true,
-  its: true,
-  ci: true,
-  jobs: true,
-  updateJobs: true,
-  export: true,
-  server: false,
-};
-const targetPath = path.resolve('.');
-ctx.setOptions(indexerOptions);
-ctx.setTargetPath(targetPath);
-conf.loadConfig(ctx);
-const config = conf.get();
+import './base.test.ts';
 
 describe('vcs', function () {
-  const db = new Db(config.arango);
+  const db = new Db(conf.get().arango);
   const gateway = new GatewayMock();
   const reporter = new ReporterMock(undefined, ['commits', 'files', 'modules']);
